el abm de registros de usuarios esta funcional.

+++ TrabajoPracticoFinal/src/main/java/ar/edu/unju/fi/repository/IRegistroServiceRepository.java
@@ -1,11 +1,9 @@
 package ar.edu.unju.fi.repository;
 
 import java.util.List;
-
 import org.springframework.data.repository.CrudRepository;
-
 import ar.edu.unju.fi.entity.Usuario;
 
-public interface IRegistroServiceRepository extends CrudRepository<Usuario, Integer> {
+public interface IRegistroServiceRepository extends CrudRepository<Usuario, Long> {
 	public List<Usuario> findByEstado(boolean estado);
 }
